+++ projects/demo-app/src/main.ts
@@ -1,0 +1,11 @@
+import { enableProdMode } from '@angular/core';
+import { environment } from './environments/environment';
+import { AppComponent } from './app/app.component';
+import { bootstrapApplication } from '@angular/platform-browser';
+
+if (environment.production) {
+  enableProdMode();
+}
+
+bootstrapApplication(AppComponent)
+  .catch(err => console.log(err));